Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> I think that we need to go back to either having a sequence number or, probably better, tracking the monotonic timestamp of each entry.  Just relying on the realtime time stamp is not enough, since there are plenty of bad cases we can get into.
> 
> I'm just not sure if it's better to track the monotonic timestamp of when the entry reaches logd, which would be a slightly better sequence number, or if we should take a timestamp supplied by the client, which would be more accurate, but could be manipulated by clients to insert log messages into arbitrary positions in the log, causing a bunch of potential issues.  For the safety reason, I think the first is the best.

When I try to use monotonic time. And the recorded time will running after the booting time, so there will be no problem that the original timestamp is adjusted， and the logd memory leak will not happen.

Whether we could do like this way, use the monotonic (currently real time) timestamp to record. In order to show the natural time, we could add a field in LogBufferElement to record realtime. When we deleting the log, it could follow the monotonic time, and the nature realtime will be recorded in log files.

This patch have been verified, and  the corresponding CTS test is also passed. There is one influence that will increase about 300kBytes as each log add one field.

Patch-set: 2
Reviewer: Gerrit User 1129218 <1129218@85c56323-6fa9-3386-8a01-6480fb634889>