Update patch set 13
Patch Set 13:

> Patch Set 12: Code-Review+2
> 
> (1 comment)
> 
> lgtm.  I'm still wondering if sspatil@ has any comments about using the kernel command line or exposing this in any case.

Actually, in this case its probably a legitimate use of the cmdline since we want the Init process to stop. The problems I usually have with are to do with bootloader passing messages to the Android (not init) via kernel commandline.

(I still think "androidboot.xxxx" is wasting too many precious characters, but this won't be in the production anyway, so we are ok)

Patch-set: 13
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>