Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> Just wondering what's the status with this.  I was making some changes to init today and expected to see some of this code and noticed that it wasn't there.

Unfortunately we got more pushback for getting this into Q, so that's not going to happen. Given the proposed linker work, this is not going to be relevant for R, so I think we can drop this. I'm sorry that you spent much time reviewing this. Although I do hope that you agree the refactoring we made for this are still better than nothing. If you'd like me to revert those, let me know.

Patch-set: 15
Reviewer: Gerrit User 1349653 <1349653@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0