Update patch set 3
Patch Set 3:

Replying only as a historian.

> I was assuming that we only had one thread writing the log messages, which makes the monotonic clock also a sequence number, but I now realize that the audit log and the kernel log come from different threads, so that may also create a situation where we have out of order timestamps.

And... each of these threads can prune content, not just at the oldest end of the list.

I regret saving memory space and one time lookup for performance by removing the monotonic clock and relying on just the realtime clock with the hopes that in-place sorting would look after this. There was an interim time(sic) we had a pure sequence number to prevent the time lookup, but it was not enough performance saving. However, for functionality, we really need to support placing Realtime, Monotonic_raw (because this is what kernel actually uses for dmesg) and Boottime values into the logs; but doing so costs us three time lookups/per.

> I've been thinking of changing flushTo() to take an return iterators instead.

Prune can cause an iterator to go away, and with heavy load from the spam filter from a large swath in the middle ... already tried (something like) that and got the surprise of a lifetime. Maybe we should invent a 'safe iterators' class that can get notification of context loss and skips to the next entry?

Patch-set: 3
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>