Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> did you get the real numbers from static linking yet?

I couldn't check all APEXes but conscrypt APEX.
From conscrypt APEX I got following numbers.

For now, it contains following so files and total size is 4,650,240.
lib64/libssl.so	355,728
lib64/libc++.so	830,592
lib64/libjavacrypto.so	292,032
lib64/libcrypto.so	1,244,712
lib/libssl.so	242,824
lib/libc++.so	574,080
lib/libjavacrypto.so	222,584
lib/libcrypto.so	887,688

If libjavascypto.so is statically linked to libc++_static, total = 3,698,560
lib64/libssl.so	355,728
lib64/libjavacrypto.so	560,848
lib64/libcrypto.so	1,244,712
lib/libssl.so	242,824
lib/libjavacrypto.so	406,760
lib/libcrypto.so	887,688

If libjavacrypto.so is statically linked to all dependencies, total = 3,372,868
lib64/libjavacrypto.so	1,956,808
lib/libjavacrypto.so	1,416,060

If the conscrypt APEX can omit libc++.so by relying on the common APEX, total = 3,245,568
lib64/libssl.so	355,728
lib64/libjavacrypto.so	292,032
lib64/libcrypto.so	1,244,712
lib/libssl.so	242,824
lib/libjavacrypto.so	222,584
lib/libcrypto.so	887,688

If libjavacrypto.so is statically linked to dependencies but c++(relying on the common APEX), total = 2,933,308
lib64/libjavacrypto.so	1,692,336
lib/libjavacrypto.so	1,240,972

Patch-set: 1
Reviewer: Gerrit User 1415877 <1415877@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0