Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Since we are going LMKD diet to make it fast, this CL goes opposite way, IMO.

100% agree, but pedantic correctness _always_ seems to be that cost ;-(

> If there is something we should read 4KB above, that gives me a signal "we should fix something", rather than allowing it.

The "fix something" is the kernel which has a different velocity than lmkd. We may not have the ability to signal a partner kernel developer, taking a patch, not taking a patch ... (what is it, 13000 different hardware platforms last year?)

. . . questioning feature . . .

> It would be very rare to change on demand(Actually, with my short experience, I never have seen to change it in runtime).

In my _long_ experience, so many bugs in how procfs works, this was listed as the primary reason at the beginning of the century to deprecate procfs and to create sysfs interface (which limits nodes to one value each, and 4KB tops).

> If someone want to really change it on runtime, can't we create some API so that LMKD only get the message when it is updated. With that, LMKD doesn't need to re-read the same value all the time.

inotify would be nice, but not available unless the signal is moved to where uevent is available?

> About the status, it is used by only to get tgid. Then, are't we enough to read 1Kbyte from the status?

(I agree pragmatically hence why I have not 'voted' on this CL, I am happy that it is present to measure intent, and to languish until we find we really need it since lmkd will 'still work' albeit w/o this feature)

Patch-set: 11
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>