Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Since we are going LMKD diet to make it fast, this CL goes opposite way, IMO.
> If there is something we should read 4KB above, that gives me a signal "we should fix something", rather than allowing it.

Yes, having to read more than 4KB *is* a signal that something is wrong. What's wrong is the text-based procfs interface that forces us to do this reading in the first place! I'd be all for adding an efficient binary interface for getting this sort of information out of the kernel, and I've made specific proposals along these lines. See [1]. *In the meantime* we need to use procfs correctly, and "correctly" here means writing code with the understanding that the kernel's API contract with userspace doesn't include putting the information we want on the first 4KB of the files that we read.

If the >4KB read thing bugs you, work with me on adding a better binary interface.

[1] https://lore.kernel.org/lkml/CAKOZuetx5QDPBEcUT4A2vNq3rbN_2SRMDVMeYP5p+didtUV0bA@mail.gmail.com/

Patch-set: 11
Reviewer: Gerrit User 1130394 <1130394@85c56323-6fa9-3386-8a01-6480fb634889>