Update patch set 12
Patch Set 12:

> Patch Set 11:
> 
> Since we are going LMKD diet to make it fast, this CL goes opposite way, IMO.
> If there is something we should read 4KB above, that gives me a signal "we should fix something", rather than allowing it.
> 
> zoneinfo, kernel already went node LRU instead of zone's one.
> IOW, zoneinfo would be less important than old day by day.
> Why do we need to read zoneinfo?
> Only reason at this moment is get to totalreserve_pages which is add up of protection pages of each zone.
> They are almost one time setup values at booting and updated with lowmem_reserve_ratio.
> It would be very rare to change on demand(Actually, with my short experience, I never have seen to change it in runtime).
> If someone want to really change it on runtime, can't we create some API so that LMKD only get the message when it is updated. With that, LMKD doesn't need to re-read the same value all the time.
> 
> About the status, it is used by only to get tgid. Then, are't we enough to read 1Kbyte from the status?

I do agree that reading and parsing large files is not the best way for finding out how hard we are reclaiming or Tgid of a process but at the moment that's the only way we have. If this whole patchset proves to be useful in testing and we adopt this new kill strategy then we can follow up with optimizations on how this information can be extracted from the kernel more efficiently. However I can't jump two steps ahead or wait for this information to be presented without a concrete proof that it is useful and absolutely necessary. IMHO this is an incremental step in the right direction. As for reading /proc files partially, I think it's a slippery slope. I've seen for example zoneinfo format slightly differ from vendor to vendor, so would not want to make any assumptions as of where a specific field can be found.

Patch-set: 12
Reviewer: Gerrit User 1157738 <1157738@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0