Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > At this point it may also make sense to explicit the SnapshotState enumerator values and use the enum in the state file.
> > > I don't see any added value in having a human-readable string in the state file.
> > 
> > The reason I didn't want to do this is then we have to worry about someone changing an enum value. Assigning fixed values and using static asserts could address that, I guess.
> 
> Yes, that's what I meant.
> @Yifan, what do you think?

I disagree. I think having a human-readable string is good because it is explicit about the merge state. The number (index) of a state is more of an implementation detail of libsnapshot. For example, you'll be able to assign the states with different numbers, or reorder the states (although, I agree that's kinda useless).

Also, we are going to put /metadata in a bugreport, and a human-readable string helps.

Patch-set: 1
Reviewer: Gerrit User 1121838 <1121838@85c56323-6fa9-3386-8a01-6480fb634889>