Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > At this point it may also make sense to explicit the SnapshotState enumerator values and use the enum in the state file.
> > > I don't see any added value in having a human-readable string in the state file.
> > 
> > The reason I didn't want to do this is then we have to worry about someone changing an enum value. Assigning fixed values and using static asserts could address that, I guess.
> 
> Yes, that's what I meant.
> @Yifan, what do you think?

(It's the same argument when you want to serialize an enum to a JSON string. An integer is less stable and may fail silently if the client and server have a different set of ordinals. A string is more robust because clients / servers can handle unknown strings explicitly).

Patch-set: 1
Reviewer: Gerrit User 1121838 <1121838@85c56323-6fa9-3386-8a01-6480fb634889>