Update patch set 1
Patch Set 1: Code-Review-2

This is intentional, it ensures that DAP devices have kept the super partition compatible with AOSP tooling. I'm fine with moving it into a new test (and adding it to VTS as well), but we shouldn't remove it without doing that first.

Patch-set: 1
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2