Update patch set 1
Patch Set 1:

That also explains why you are not seeing the FIPS crash in init_first_stage - this builds fine on taimen (which should be the same as walleye), which means there are no references to libcrypto.a and it will be stripped when linking, including the FIPS self test.  It doesn't build on cuttlefish, which is what I was testing with.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0