Update patch set 2
Patch Set 2:

> Patch Set 2: Verified-1 Presubmit-Ready+1 Code-Review-1
> 
> (1 comment)
> 
> The picture is not complete, the SIGHUP from executing fsck android_fork_execvp_ext was a surprising wild-card. Was not an issue before in first stage init, but now is because we will _find_ /system/bin/fsck.* dealing with fsck of scratch. I have not tested that scenario.

We do find /system/bin/fsck.* in first stage init too.  We mount system and do the switchroot before mounting the rest of the partitions, so they can be fsck'ed.

Something must be different in order for us to get SIGHUP in this case.  Perhaps it's because we're running in a shell through adb and that's changing something?

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>