Update patch set 3
Patch Set 3: Code-Review-1

Not that this CL will always change the vbmeta flags in boot.img (no matter it is the top-level vbmeta struct or not)

So when boot.img is chained, i.e., there is a chain descriptor for "boot" in vbmeta.img,
the chained vbmeta at the end of boot.img will be changed as well.

On next boot, libavb will refuse to boot:
https://android.googlesource.com/platform/external/avb/+/refs/heads/android10-dev/libavb/avb_slot_verify.c#762

However, I don't have a good solution for this yet.
Probably in fastboot, tracking whether there is vbmeta.img or not. (i.e., if no vbmeta.img and --disable-verification is provided, rewrite boot.img's vbmeta).

Patch-set: 3
Reviewer: Gerrit User 1080813 <1080813@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1