Update patch set 1
Patch Set 1:

By itself, this is a necessary cleanup, so push it for that reason alone.

[TL;DR]

I must admit that this was in my internal patch set searching for a solution, and even the NOCTTY in the child too, and it did not solve the problem for me. But when you posted this I thought you did it right and I did it wrong when I first saw this (it happens ;-) ). For clarity, are you saying it did not work?

I will be gone for the remainder of the week, regardless if this works or not, please merge it as a good cleanup that correctly opens the child_ptty only in the context of the child. And if it does not work, just be less forceful about the commit message...

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>