Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Indeed not having the sigmask get reset is the issue.  It seems like that line is just forgotten, since presumably it should be done as long as ignore_int_quit isn't set.
> 
> Seems even more so that we can get rid of all the sigaction() lines too.  Just having that sigmask should be enough to ignore those signals.
> 
> I'm not even super sure that we'd want to ever ignore them though.  What good does it do us by ignoring those signals?  It doesn't seem to work at TOT either, since `adb shell` followed by `logwrapper /system/bin/yes` still exits for Ctrl-c.

I chatted with enh@ today about this.  We both figure that this shouldn't even be an option.  The expected behavior should always be to handle SIGINT and SIGQUIT by passing it to the child process.  This is only a wrapper after all.

> 
> Also, looking at the history, the child_ptty opening was moved into the parent in ag/286794.  The reasoning is that the child_ptty must be opened for POLLHUP to be generated, which is fair.  It seems we'd need a new mechanism for detecting when the child died.

Same conversation: we figure that the most straight-forward way to fix this is to fork another process here that does the actual logwrap behavior.  This way we can use a signal handler for SIGCHLD (or better: signalfd) like we did originally but without the problems of changing the parent's signals.

I'll push something tomorrow.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>