Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I was figuring that we would only want to opt-in executables that cannot have normal Android logging to this, particularly those that are run early in the boot process, since we should prefer later services sending their logs to logd.
> 
> i think that was the _theory_, but i also think it was unrealistic to assume "everyone will just rewrite all their stuff". sure, you _can_ rewrite shell scripts to use logger instead of echo/printf, but it's extra work and less portable. you can't really do much with toybox tools you call. you _can_ use Android logging for code you've written yourself ... but as soon as you call out to a third-party library or whatever, you're going to lose that logging.

I was wondering if we'd want to direct logs to logd for services that start later.  Though, I guess, there should be few enough services writing to stdout/stderr that it wouldn't be too bad to always put it in kmsg.  Having the linker messages always in kmsg would be a plus (and they already go to logcat for the later services)

> apps have this problem too, especially with third-party libraries. that's why i think we should have the environment variable to redirect stdio (though apps would still need to use wrap.sh, which is why i've never actually _done_ anything other than mumble about how we should do something better!).

I was thinking about this a bit - the idea of moving the logwrapper logic into libc.  We don't want to rely on logd, since most of our logwrapper cases right now are in early init.  We could have a pipe from stdout/stderr to a separate thread that we spawn that does the logwrap actions?  It'd be difficult to capture the linker errors too if we did this, as we'd need to put that behavior even earlier.

Though the app case is a bit tangent from there too.

Patch-set: 1