Update patch set 3
Patch Set 3: Presubmit-Ready+1

I went though a bunch of the history today and was looking into trying other options.  

Since this is a library function and we don't want to mess with signals in any case except `logwrapper` itself, POLLHUP is the best option.  We don't have pidfd yet, so the timeout workaround is likely the best option if the child happens to exit before it opens child_ptty.

The ignore_int_quit code has been broken for a long while.  It also doesn't make sense to have in a library function.  It doesn't even make sense to have in `logwrapper`, since we have `nohup` now, so that's removed.

I do add code to forward the signals to fix a test case that is pretty rough and related to the above.  It still needs work though as there is a race when starting the child process and again when reaping it.

This should be much better than before though.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1