Update patch set 3
Patch Set 3:

> I agree with all that nnk@ says above.  I think it's fine to make this only an opt-in, especially with the goal of limiting its usage.

I gathered from Nicks response in [1] that making it the default in eng and userdebug builds would be fine, right?

I'd really rather ensure there's log coverage by default (just in local and test builds is good enough), because a) it covers integration-level cases where it's otherwise hard to both anticipate the problem and fix it locally in one component (the linker is a prime example, but I wouldn't assume it's the only one), and b) it sucks a lot of engineering time to dig into failures when logs go to /dev/null. We should assume it will often hit engineers that don't know the boot process intimately, and first wouldn't suspect that to be the case and then would have to spend time figuring out what to do about it.

> With regards to the linker, I don't have a great option here.  Perhaps we can append more information to the log when the exec() call fails to state if the error is "No such file or directory", then it could be a linker error and the linker should be debugged directly?

Yeah, drawing some attention to the linker could help, but for root causing the dlerror message is key and it'd save a lot of time to see it right away. For further linker debugging stdio_to_kmsg would do, and in particular the fallback in liblog to use the kernel log before logd is ready, as discussed in [2] (which btw probably requires sepolicy changes to work).

> Truthfully, before your bug I've only seen one other case of a linker failure so early in boot.

Fair enough, but I fear there's more to come - a lot of work is going on both in APEXes that affect boot and in linker config handling for APEXes.

[1] https://r.android.com/c/platform/system/core/+/1127617/3#message-751cef166eccf01664cce07de522f11e6e276714
[2] https://r.android.com/c/platform/system/core/+/1119215/2#message-02f79e15cc691fe292e0ccaa4c3d6db1d5f7133f

Patch-set: 3
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>