Update patch set 4
Patch Set 4: Verified+1 Presubmit-Ready+1

This is ready for review.  

I tested pretty much every case I could think of and I'm confident in this now.  It doesn't do anything with signals when called without forward_signals, which is the right case for using it as a library.  It does do the right thing and forward these signals to the child process when called in logwrapper.

I definitely learnt a lot about TTY's in this CL.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1