Update patch set 3
Patch Set 3:

> Sounds like a plan to me.  I filed b/141617891 for sending stdout/stderr to logd.

Where would that be implemented; presumably not in init then? Wouldn't it have both the same benefits and drawbacks as has been discussed here? (Besides, I think capturing stderr should be enough - there are valid scenarios for sticking to stderr logging to limit dependencies, but that doesn't apply to stdout.)

As for the linker specifically, I think the most important AI is to make liblog fall back to kmsg before logd is ready. That would cover both the initial "CANNOT LINK" message with the dlerror, and more verbose debug we might need to enable. I gathered the intention is to fix that as part of http://b/119867234.

Patch-set: 3
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>