Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Sounds like a plan to me.  I filed b/141617891 for sending stdout/stderr to logd.
> 
> Where would that be implemented; presumably not in init then? Wouldn't it have both the same benefits and drawbacks as has been discussed here? (Besides, I think capturing stderr should be enough - there are valid scenarios for sticking to stderr logging to limit dependencies, but that doesn't apply to stdout.)

Right, it would be implemented in libc.  It is tangent with what you're interested in solving for the linker.

> 
> As for the linker specifically, I think the most important AI is to make liblog fall back to kmsg before logd is ready. That would cover both the initial "CANNOT LINK" message with the dlerror, and more verbose debug we might need to enable. I gathered the intention is to fix that as part of http://b/119867234.

The linker does not use liblog, it directly writes to stderr and the logd socket.  I don't think that the problem is that we need a fallback to kmsg when logd isn't available, since logd is available extremely early, so that isn't a good enough signal.

I think the real problem here is "if the system doesn't boot, can we still get the relevant logs from logd".  Admittedly this is something we need to solve, but I don't have a quick solution for it.  I filed b/141634082 to keep this in mind.

I still wonder, how many other people have been burnt by linker failures at such a very early stage.  And if having the log when exec fails in init be explicit with debugging instruction would be enough to have unblocked them.  I believe this would be sufficient, especially since builds that are failing very early (before adb launches), already require a decent amount of hands on debugging.

Patch-set: 3