Update patch set 5
Patch Set 5: Code-Review+1

Actually, I wanted to make it based of function of "troubleness". IOW, if we see the thrashing for a sec, it just kill background apps. But if we keep seeing the thrashing for several sec(e.g., 5sec), we could lower the threshold proportionally so that it could kill finally even top app if we see thrashing for a long time(e.g., 10sec).
However, don't have enough time to test/implement from my side so I'm not objecting it if Suren want to go with this.

Patch-set: 5
Reviewer: Gerrit User 1275554 <1275554@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1