Update patch set 1
Patch Set 1:

(2 comments)

I think this can be even more helpful if it is implemented more centrally in libmodprobe. With the check in Insmod() we'll still end up iterating over a module's soft and hard dependencies in the library, before terminating in Insmod (assuming the module is loaded). If the check is put at the top of LoadWithAliases we can prevent that also.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>