Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> However, the boards don't have batteries, so maybe we're missing some component between the hals which is why this hasn't been seen elsewhere?

So on the HiKey960 it looks like I can solve this by fixing an sepolicy so that it sees a power_supply in sysfs. However, on the db845c the kernel currently doesn't export a power_supply so we still see the issue there.

So we can probably work around it, but it does seem that this change has somehow changed the behavior in the no-battery/no-power-supply case. Previously it would stay running, and now it suspends fairly quickly (which seems a little odd for a no battery situation)/

Patch-set: 5