Update patch set 3
Patch Set 3:

You don't need this.  Maybe we can just replace this CL with better comments on what /dev/.booting is used for.

Back before Treble, some kernel firmwares that ueventd loads were on the ramdisk and others were on partitions loaded during the mount_all in `on fs`.  This was a hack that someone put in to keep retrying to load the firmware until we've mounted all the partitions.  We use removing this file as an indication that they've been mounted.  

Ideally with Treble, all partitions that contain firmwares should be mounted during first stage mount, so this check should be not needed.  Unfortunately, some pixel devices still mount their modem firmwares during `on fs`, so it is needed.  

We don't load firmwares directly from /data, so there's no need to recreate this file then remove it later.  That would just make the firmware loaders keep trying while /data is unmounted, which isn't needed, since we aren't unmounting any of the partitions from where we load firmwares.

Full disclosure: it looks like some new devices do copy firmwares from /data to a tmpfs that they use to load the firmwares, to work around us not wanting to load firmwares from /data...  But that doesn't change anything above, since once the files are copied they can be loaded regardless of the state of /data.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>