Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> i thought the problem with bpfmt was inconsistency because there isn't a prebuilt that's used... did that get fixed?

This is a separate problem, maybe which I wouldn't even consider a problem. bpfmt should be deterministic, and if we change the format, then well, we change the format. When you use -s in a configuration like this though:
https://googleplex-android.git.corp.google.com/platform/system/tools/hidl/+/73bf8b24f1a3d1a1658dea5228c06641942490d4/PREUPLOAD.cfg#5

preupload hook won't tell you that it's running this command, so it might complain, and 'bpfmt -w ...' alone won't be enough to fix it. This is why I'm not following up w/ a change to the preupload hook.

Patch-set: 3
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>