Update patch set 2
Patch Set 2: Presubmit-Ready+1

(3 comments)

> Patch Set 1:
> 
> Is this needed when using the new CreateDevice signature? It accepts a timeout parameter and guarantees the returned device hasn't raced.

In case of apexd, we create and destroy dm-devices during pre-reboot verification. Synchronizing on them being deleted makes code cleaner and easier to test. Before this change apexd had it's own implementation for this synchronization which had a bug (apexd was waiting for /dev/block/dm-XX to be deleted, not the /dev/block/mapper/by-uuid/<uuid>). Since apexd most likely is not the only process that is interested in waiting for dm-devices to be deleted, I've decided that it makes sense to have correct implementation of the synchronization in libdm, instead of all the clients implementing their own (most likely wrong) logic. But if you think that it's not useful to have in libdm, I'm happy to revert the change and implement it in apexd.

Patch-set: 2
Reviewer: Gerrit User 1432579 <1432579@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1