Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Something seems off to me still, but I can't exactly put my finger on it.
> 
> What happens on the main thread vs other threads?  It looks like both the promotion from weak_ptr to a normal pointer and the deletion of the real object are both on the main thread?  And we are fully consuming the atransport in the main thread too, right?  I don't see a ref count, so I assume so.

The users of atransport* aren't necessarily on the main thread, but they're the ones that submit the work to delete the atransport to the main thread, so that should be fine.

> What if we did this instead?
> 
> std::set<atransport*> valid_atransports;
> bool check_atransport(atransport* ptr) {
>   // Check main thread
>   // return true if in valid_atransports
> }
> void delete_on_main_thread(atransport* ptr) {
>   valid_atransports.remove(ptr);
>   delete ptr;
> }
> 
> Or if you don't want to store the pointer for risk of misusing it, perhaps keep the map and the 'id', but have the same concept, where the main thread just removes the entry in the map when it deletes the object.
> 
> What doesn't work about that, that requires the weak_ptr<> class?

Isn't that basically the same thing? I prefer weak_ptr because the purpose/validity is clearer: having a raw pointer or id floating around that's tied to an entry in a list somewhere seems worse to me.

Patch-set: 3
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0