Update patch set 4
Patch Set 4:

> > Isn't that basically the same thing? I prefer weak_ptr because the purpose/validity is clearer: having a raw pointer or id floating around that's tied to an entry in a list somewhere seems worse to me.
> 
> If this is the same, then I think we can remove a bit of the weak_ptr<> class and we don't need the std::vector in enable_weak_from_this<>, right?  I don't see where that std::vector is needed.  We could have 
> 
> void schedule_deletion() {
>         fdevent_run_on_main_thread([this]() { 
>           // remove the entry from the transports map
>           delete this; 
>         });
>     }
> 
> And skip a lot of the extra code?

It's less code, but it feels a lot more unnatural to me. Having a global list of objects of type T is distasteful (and also is probably more prone to bugs, since the invalidation of weak_ptr happens in schedule_deletion instead of the destructor: an accidental call to `delete foo` instead of `foo->schedule_deletion()` will leave weak_ptrs alive.

> Alternatively to the above we could do:
> 
> template <typename T>
> class weak_ptr {
>   weak_ptr(T* t) : t_(t) {}
>   T* promote() {
>     // Check main thread
>     if (weak_ptr_source<T>::valid_ts.contains(t_)) { 
>       return t_; 
>     }
>     return nullptr;
>   }
> };
> 
> template <typename T>
> class weak_ptr_source {
>  public:
>   weak_ptr_source() {
>     valid_ts.emplace(this);
>   }
>   ~weak_ptr_source() {
>     valid_ts.erase(this);
>   }
>   weak_ptr<T> get_weak_ptr() {
>     return weak_ptr(this);
>   }
>  private:
>   static std::set<T*> valid_ts;
> };
>
> We'd need construction of weak_ptr_source to happen on the main thread.  Is that true?  Otherwise, is get_weak_ptr() called only on the main thread?  We could do the emplace there instead.
> 
> Then you'd use atransport->weak_ptr() as the argument.  Then users would call promote() on it like a normal weak_ptr.  It'd be less overhead than the map, get rid of the 'id' that gets passed as a parameter and doesn't need the non-static std::vector.  Does it take care of everything or am I missing something?

Reading this proposal more carefully, I realized why I had an instinctive dislike of the previous proposal: they're both broken, because a weak_ptr can go from valid to invalid back to valid, if the allocator puts a new object of that type at the same address. We *have* to immediately invalidate all of the weak_ptrs created from an object when it's destructed, and the only way to do that is to maintain a list of all of the weak_ptrs that point to it.

Patch-set: 4
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0