Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Actually... Jiyong, is it OK to allow static linking of this library? It has:
> 
>     vendor_available: true,
>     vndk: {
>         enabled: true,
>     },

That's not desirable. We don't want vendor to be able to statically link to VNDK libraries. That's against the goal of VNDK. But since you are statically linking to it only from the test, it is not an immediate problem.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>