Update patch set 2
Patch Set 2:

So, is the idea here that we don't trust userspace reboots? So we give it a free attempt at a userspace reboot which we don't count against the attempts, and the next reboot is always guaranteed to be a hard reboot? Because if we can get another userspace reboot from this, wouldn't we loop until something else intervenes?

Patch-set: 2
Reviewer: Gerrit User 1045980 <1045980@85c56323-6fa9-3386-8a01-6480fb634889>