Update patch set 9
Patch Set 9:

This doesn't make sense to me.  RemoveProcessGroup() is only called in KillProcessGroup() when it has iterated through the cgroup and found no processes remaining: https://cs.android.com/android/platform/superproject/+/master:system/core/libprocessgroup/processgroup.cpp;l=359?q=removeprocessgroup

It's impossible for it to fail for the stated reason that process death takes time, since we're already waiting for the directory to be completely empty in KillProcessGroup(), which already has a sleep and retry loop.

If this is actually reproducing, then it needs more investigation; particularly understanding what is happening between when DoKillProcessGroupOnce() returns 0, indicating that the process group is empty, and when RemoveProcessGroup() fails due to seeing a non-empty directory.

Patch-set: 9
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>